Deleting old API in preparation for v0.4.0
